add sonar analyzer and refactoring

+++ Ecommerce.Application/ProductCategories/DeleteProductCategory/DeleteProductCategoryCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using Ecommerce.Application.Abstractions.Messaging;
 using Ecommerce.Domain.Abstractions;
 using Ecommerce.Domain.ProductCategories;
+using Ecommerce.Domain.Products;
 
 namespace Ecommerce.Application.ProductCategories.DeleteProductCategory;
 internal sealed class DeleteProductCategoryCommandHandler : ICommandHandler<DeleteProductCategoryCommand, Guid>
@@ -16,11 +17,16 @@ internal sealed class DeleteProductCategoryCommandHandler : ICommandHandler<Dele
 
     public async Task<Result<Guid>> Handle(DeleteProductCategoryCommand request, CancellationToken cancellationToken)
     {
-        var productCategory = await _productCategoryRepository.GetByIdAsync(new ProductCategoryId(request.id));
+        ProductCategory? productCategory = await _productCategoryRepository.GetByIdAsync(new ProductCategoryId(request.Id), cancellationToken);
+
+        if (productCategory is null)
+        {
+            return Result.Failure<Guid>(ProductCategoryErrors.NotFound);
+        }
 
         _productCategoryRepository.Remove(productCategory);
 
-        await _unitOfWork.SaveChangesAsync();
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
 
         return productCategory.Id.Value;
